requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
httpx>=0.25.0
tiktoken>=0.5.0
sqlite-vec>=0.1.0
FlagEmbedding>=1.2.0
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
mcp>=1.0.0
beautifulsoup4>=4.12.0
